first step to objects building

+++ plugins/core/landscape_model/sources/landscape/lm_landscape.cpp
@@ -197,7 +197,7 @@ Landscape::getObject( const Object::UniqueId& _id ) const
 int
 Landscape::getObjectsCount() const
 {
-	return m_objects.size();
+	return static_cast< int >( m_objects.size() );
 
 } // Landscape::getObjectsCount
